# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  floorplan.FloorView:
    properties:
      activePlanId:
        description: |-
          ActivePlanID names the plan PlanURL was built from, so a client that
          already has that plan's id (e.g. from the admin history listing) can
          tell it is the active one directly, rather than reconstructing the
          fact by checking whether its digest appears inside the URL.
        type: string
      buildingId:
        type: string
      code:
        type: string
      floorId:
        type: string
      levelOrder:
        type: integer
      name:
        type: string
      planUrl:
        type: string
      viewBox:
        type: string
    type: object
  floorplan.Stored:
    properties:
      createdAt:
        type: string
      createdBy:
        description: |-
          CreatedBy is nil for the plans seeded from packages/floorplans, which
          no admin uploaded.
        type: string
      floorId:
        type: string
      planId:
        type: string
      sha256:
        type: string
      viewBox:
        type: string
      warnings:
        items:
          $ref: '#/definitions/floorplan.Warning'
        type: array
    type: object
  floorplan.Warning:
    properties:
      code:
        description: |-
          Code is the stable machine-readable reason; display text is the
          client's (ADR-0012).
        type: string
      ref:
        description: Ref is the place id, node id, or element the warning is about.
        type: string
    type: object
...
  httpx.ErrorResponse:
    properties:
      code:
        example: not_found
        type: string
      error:
        example: visit not found
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Patient journey and indoor navigation API above the HIS.
    Serves the CarePath-derived journey plan (ADR-0009) with
    every actionable step resolved to its service point.
  title: CarePath API
  version: 1.0.0
paths:
...
  /api/v1/admin/floors/{floorId}/plan:
    post:
      consumes:
      - image/svg+xml
      description: 'FR-11/ADR-0015: replaces the floor''s plan. Send the SVG as the
        raw request body. The upload is parsed, held to a closed allowlist and re-serialized
        before storage, so a rejection means the file has to change — nothing is silently
        stripped. A place or navigation node the plan does not draw is a warning rather
        than a rejection: the app already shows such a place as not routable yet.
        Plans are append-only, so this adds a row and moves the floor''s pointer;
        re-uploading an unchanged file is idempotent.'
      parameters:
      - description: Floor id (e.g. I-1301)
        in: path
        name: floorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/floorplan.Stored'
        "400":
          description: the plan broke a rule — the message names which
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: missing, malformed, or expired staff access token
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: authenticated, but not an ADMIN
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: unknown floor
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Upload a floor plan
      tags:
      - admin
  /api/v1/admin/floors/{floorId}/plans:
    get:
      description: 'ADR-0015: every plan stored for this floor, newest first, with
        the warnings each was accepted with. Plans are never deleted, so this is both
        the history and the list a rollback picks from.'
      parameters:
      - description: Floor id
        in: path
        name: floorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/floorplan.Stored'
            type: array
        "401":
          description: missing, malformed, or expired staff access token
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: authenticated, but not an ADMIN
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: unknown floor
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - bearerAuth: []
      summary: List a floor's plans
      tags:
      - admin
  /api/v1/admin/floors/{floorId}/plans/{planId}/activate:
    post:
      description: 'ADR-0015: the rollback. Plans are append-only, so returning to
        an earlier drawing moves the floor''s pointer rather than restoring anything.'
      parameters:
      - description: Floor id
        in: path
        name: floorId
        required: true
        type: string
      - description: Plan id from the plans listing
        in: path
        name: planId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/floorplan.Stored'
        "401":
          description: missing, malformed, or expired staff access token
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: authenticated, but not an ADMIN
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: no such plan on that floor
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Make a stored plan the floor's active one
      tags:
      - admin
  /api/v1/admin/floors/{floorId}/plans/{planId}/raw:
    get:
      description: 'ADR-0015: the original bytes, before normalization, so an admin
        can get back the file they sent. This is never what a patient is served —
        that is always the normalized artifact.'
      parameters:
      - description: Floor id
        in: path
        name: floorId
        required: true
        type: string
      - description: Plan id from the plans listing
        in: path
        name: planId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: the uploaded file
          schema:
            type: string
        "401":
          description: missing, malformed, or expired staff access token
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: authenticated, but not an ADMIN
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: no such plan on that floor
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Download a plan as it was uploaded
      tags:
      - admin
...
  /api/v1/floors:
    get:
      description: '#105/ADR-0015: every floor in display order with the URL of its
        current plan. This is the pointer document the web app revalidates; the plan
        URL it hands back is immutable and cached for a year. Replaces the floor list
        apps/web used to hardcode beside its bundled plan assets.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/floorplan.FloorView'
            type: array
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List floors and their plan URLs
      tags:
      - floors
  /api/v1/floors/{floorId}/plan/{asset}:
    get:
      description: '#105/ADR-0015: the normalized SVG for a floor, addressed by its
        own sha256. The response is immutable — a different drawing is a different
        digest and therefore a different URL — so it is cached for a year and never
        revalidated. Take this URL from the floors listing rather than assembling
        it.'
      parameters:
      - description: Floor id (e.g. I-1301)
        in: path
        name: floorId
        required: true
        type: string
      - description: The plan's sha256 with an .svg suffix, as the floors listing
          spells it
        in: path
        name: asset
        required: true
        type: string
      produces:
      - image/svg+xml
      responses:
        "200":
          description: the floor plan SVG
          schema:
            type: string
        "404":
          description: no plan with that digest on that floor
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get a floor plan
      tags:
      - floors
...
swagger: "2.0"